chore(lib): renamed webgl1.ts -> webgl.ts
